+++ HISTORY.md
@@ -1,5 +1,9 @@
 History of features added to LiChess Tools in time (reversed order)
 
+13 Jan
+
+- hide individual side elements in lobby
+
 12 Jan
 
 - more ceval lines